Adding file "source\ScriptModules\ScriptModulesOptional.h"
Reason, for easier overview.
