+++ AutoHotkeyx.vcxproj.filters
@@ -225,6 +225,9 @@
     <ClInclude Include="source\ScriptModules\ScriptModulesNested.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="source\ScriptModules\ScriptModulesOptional.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="source\resources\icon_filetype.ico">
